Added html5 storage Non webkit browsers now delete users upon closing or refreshing window Get connected users Gets new messages each 2.5 seconds Only messages written after user connects are shown

+++ files/setuser.php
@@ -13,13 +13,14 @@
 		$query = 'INSERT INTO babbler_users(user) VALUES(\''.$username.'\');';
 		mysqli_query($db,$query)
 		or die('Error adding user do database!');
+	} else {
+		echo 0;
 	}
-	
-	if ($oldusername != "") {
+	// deletes user being used from table
+	if ($oldusername != "" && mysqli_num_rows($results) == 0) {
 		$delolduser = 'DELETE FROM babbler_users WHERE user=\''.$oldusername.'\';';
 		mysqli_query($db,$delolduser)
 		or die('Error deleting old username!');
 	}
-
 	mysqli_close($db);
 ?>
